backend/requirements.txt
# ============================================
# AI Code Reviewer — Backend Dependencies
# ============================================

# --- Core Framework ---
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.5.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0

# --- Email Validation (Pydantic EmailStr) ---
email-validator>=2.2.0,<3.0.0

# --- Database (MongoDB Atlas - async) ---
motor>=3.5.0,<4.0.0

# --- Auth (JWT + Password Hashing) ---
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.2.0,<5.0.0

# --- AI Agent (LangGraph + LangChain) ---
langgraph>=0.2.0,<1.0.0
langchain>=0.3.0,<1.0.0
langchain-openai>=0.2.0,<1.0.0

# --- Embedding (Google text-embedding-004) ---
langchain-google-genai>=2.0.0,<3.0.0

# --- Vector DB (Qdrant) ---
qdrant-client>=1.11.0,<2.0.0

# --- Rate Limiting ---
slowapi>=0.1.9,<1.0.0

# --- Structured Logging ---
structlog>=24.4.0,<26.0.0

# --- GitHub Integration ---
gitpython>=3.1.43,<4.0.0

# --- HTTP Client (async) ---
httpx>=0.27.0,<1.0.0

# --- WebSocket ---
websockets>=12.0,<14.0
